feat: Router type files were created for each model
