+++ app.js
@@ -1,7 +1,10 @@
 import express from 'express';
 import dotenv from 'dotenv';
-import router from './routes/routes.js';
-import { errorHandler, notFoundHandler } from './middleware/middleware.js';
+import gameRouter from './routes/gameRoutes.js'
+import playerRouter from './routes/playerRoutes.js'
+import cardRouter from './routes/cardRoutes.js'
+import scroreRouter from './routes/scoreRoutes.js'
+import { errorHandler, notFoundHandler } from './middlewares/middleware.js';
 
 dotenv.config();
 const app = express();
@@ -9,7 +12,10 @@ const port = process.env.PORT;
 
 app.use(express.json());
 
-app.use('/api', router);
+app.use('/api', gameRouter);
+app.use('/api', playerRouter);
+app.use('/api', cardRouter);
+app.use('/api', scroreRouter);
 
 app.use(notFoundHandler);
 app.use(errorHandler);
